+++ docs/about/whats_new.rst
@@ -12,6 +12,7 @@ v1.8.next
 - Fix broken paths in api docs. (:pull:`1277`)
 - Fix readthedocs build. (:pull:`1269`)
 - Added doc change comparison for tuple and list types with identical values (:pull:`1281`)
+- Add `dataset id` check to dataset doc resolve to prevent `uuid` returning error when `id` used in `None`  (:pull:`1287`)
 
 v1.8.7 (7 June 2022)
 ====================